orchestration/streamlit/requirements.txt
streamlit
pandas
altair
snowflake-connector-python
cryptography
python-dotenv
